Exercise7
Tranfering money from one account to the other.

The challenge was top spot the missing lock.unlock() method and implement that in the code, to avoid infinite wait (LiveLock)
++
Exercise6

Adding a status variable. As it is already thread-safe (since it is a local variable, it is stored in the thread stack and each thread has a copy, so No interference)